feat(AuthForms): Extract form schemas into schemas.ts
